refactor(user-profile): overview の hero 面を UserProfileHero に分離 (#707)
バナー / アバター / フォロー操作 / ユーザーメモ (#458) / ロール /
カスタムフィールド / 統計を window/user-profile/UserProfileHero.vue へ。
メモ保存とフォローは adapter prop で子が自律実行し、メニュー・QR の開閉
(親が portal で所有) と保存結果の user 反映は emit で親に返す。

モバイル向け @container クエリは hero の module に同居させた
(親 .profileContainer の container-type にコンポーネント境界を越えて反応する)。
